fix uint8_t to size_t comparison (issue #525)

+++ src/segment.c
@@ -658,7 +658,7 @@ static mi_segment_t* mi_segment_init(mi_segment_t* segment, size_t required, mi_
     memset((uint8_t*)segment + ofs, 0, info_size - ofs);
 
     // initialize pages info
-    for (uint8_t i = 0; i < capacity; i++) {
+    for (size_t i = 0; i < capacity; i++) {
       segment->pages[i].segment_idx = i;
       segment->pages[i].is_reset = false;
       segment->pages[i].is_committed = commit;
